Fix type mismatch

+++ tests/unit/numbers.spec.ts
@@ -23,19 +23,19 @@ import { attoAlphToFiat } from '../../src/utils/numbers'
 describe('numbers module', () => {
   describe('attoAlphToFiat', () => {
     test('when half of an Alph, return a decimal amount', () => {
-      expect(attoAlphToFiat(1e18 / 2, 1)).toBe(0.5)
+      expect(attoAlphToFiat(BigInt(1e18 / 2), 1)).toBe(0.5)
     })
     test('when 2 Alph, return a whole amount', () => {
-      expect(attoAlphToFiat(1e18 * 2, 1)).toBe(2)
+      expect(attoAlphToFiat(BigInt(1e18 * 2), 1)).toBe(2)
     })
     test('when 1 Alph, and 1 fiat dollar, return 1', () => {
-      expect(attoAlphToFiat(1e18, 1)).toBe(1)
+      expect(attoAlphToFiat(BigInt(1e18), 1)).toBe(1)
     })
     test('when negative Alph, return negative fiat', () => {
-      expect(attoAlphToFiat(-1e18, 1)).toBe(-1)
+      expect(attoAlphToFiat(-BigInt(1e18), 1)).toBe(-1)
     })
     test('when negative fiat price, return negative fiat', () => {
-      expect(attoAlphToFiat(1e18, -1)).toBe(-1)
+      expect(attoAlphToFiat(BigInt(1e18), -1)).toBe(-1)
     })
   })
 })
